+++ spec/integrations/rails_3_2/test.sh
@@ -14,7 +14,7 @@ bundle --version
 bundle install --quiet --gemfile="$DIR/Gemfile" --retry 2 --path vendor/bundle
 
 echo '-- cypress install'
-bundle exec ./bin/rails g cypress_dev:install --cypress_folder=cypress --install_cypress_with=npm
+bundle exec ./bin/rails g cypress_dev:install --cypress_folder=cypress --install_cypress_with=npm --no-install-cypress-examples
 
 echo '-- start rails server'
 # make sure the server is not running
@@ -24,7 +24,8 @@ bundle exec ./bin/rails server -p 5002 -e test &
 sleep 2 # give rails a chance to start up correctly
 
 echo '-- cypress run'
-node_modules/.bin/cypress run
+cp -fv ../cypress.json .
+node_modules/.bin/cypress run --record
 
 echo '-- stop rails server'
 kill -9 `cat tmp/pids/server.pid`