+++ spec/integrations/rails_4_2/test.sh
@@ -24,8 +24,9 @@ bundle exec ./bin/rails server -p 5002 -e test &
 sleep 2 # give rails a chance to start up correctly
 
 echo '-- cypress run'
+cp -fv ../cypress.json test/
 cd test
-yarn run cypress run
+yarn run cypress run --record
 
 echo '-- stop rails server'
 kill -9 `cat ../tmp/pids/server.pid`